fixed some bugs that were caused after moving files around
